+++ config/initializers/will_paginate.rb
@@ -1,0 +1,18 @@
+if defined?(WillPaginate)
+  ActiveSupport.on_load :active_record do
+    module WillPaginate
+      module ActiveRecord
+        module RelationMethods
+          def per(value = nil) per_page(value) end
+          def total_count() count end
+        end
+      end
+      module CollectionMethods
+        alias_method :num_pages, :total_pages
+      end
+      module PerPage
+        def paginates_per(value) self.per_page=(value) end
+      end
+    end
+  end
+end