+++ lib/tolk.rb
@@ -1,4 +1,3 @@
-require 'will_paginate'
 require 'safe_yaml/load'
 require 'tolk/config'
 require 'tolk/engine'